requirements.txt
google-cloud-vision>=3.4.4
google-generativeai>=0.3.1
Pillow>=10.1.0
natsort>=8.4.0
python-dotenv>=1.0.0
pdf2image>=1.16.3
tqdm>=4.66.1
requests>=2.31.0
pandas>=2.1.4
loguru>=0.7.2
ag2==0.7.3
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.8
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
fast-depends==2.4.12
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.79.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-cloud-vision==3.9.0
google-crc32c==1.6.0
google-genai==0.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-core==0.3.33
langchain-openai==0.3.3
langsmith==0.3.4
numpy==2.2.2
openai==1.61.0
opentelemetry-api==1.29.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyautogen==0.7.3
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests-toolbelt==1.0.0
rpds-py==0.22.3
rsa==4.9
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.8.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0